Changes towards Goby 1.7: Support to estimate read weights described in Hansen KD et al NAR 2010. See http://campagnelab.org/software/goby/tutorials/estimate-heptamer-weights/ In contrast to the initial publication, Goby supports using the weights to reweight annotation counts and transcript counts.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@23371 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/modes/AggregatePeaksByRPKMDifferenceMode.java
@@ -250,7 +250,7 @@ public class AggregatePeaksByRPKMDifferenceMode extends AbstractGobyMode {
                 for (final AnnotationRPKM annot : annots) {
                     final int start = annot.getStart();
                     final int end = annot.getEnd();
-                    final int overlapReads = algs[referenceIndex].countReadsPartiallyOverlappingWithInterval(start, end);
+                    final double overlapReads = algs[referenceIndex].countReadsPartiallyOverlappingWithInterval(start, end);
 
                     final double rpkm = calculateRPKM(overlapReads, annot.getLength(), numAlignedReadsInSample);
                     annot.rpkm += rpkm;
@@ -259,10 +259,10 @@ public class AggregatePeaksByRPKMDifferenceMode extends AbstractGobyMode {
         }
     }
 
-    private float calculateRPKM(final int readCountInt, final int annotLength, final int geneOverlapReads) {
-        final float readCount = readCountInt;
-        final float length = annotLength; // in bases
-        final float sampleReadCount = geneOverlapReads; // in read
+    private double calculateRPKM(final double readCountInt, final int annotLength, final double geneOverlapReads) {
+        final double readCount = readCountInt;
+        final double length = annotLength; // in bases
+        final double sampleReadCount = geneOverlapReads; // in read
         return readCount / (length / 1000.0f) / (sampleReadCount / 1E6f);
     }
